Extract GUI launch boundary
